Command line flags: disable HTTPS, change server
Sometimes, for local QA reasons, TLS cannot work. For example, when
running a Rails app locally it's a bit too much of a pain to set up a
full TLS setup. So, disable TLS with the `-0` flag.

This adds a `-s` flag, allowing the user to change the server. Useful
for QA, but also useful for interacting with competing services.